Use Pydantic time-step calculator in build_sim.py.
Move time-step calculator tests in build_sim_test.py into pydantic_model_test.py.

PiperOrigin-RevId: 736885857
