+++ torax/config/build_sim.py
@@ -15,18 +15,14 @@
 """Functions to build sim.Sim objects, which are used to run TORAX."""
 
 from collections.abc import Mapping
-import copy
 from typing import Any
-
 from torax import sim as sim_lib
 from torax.config import runtime_params as runtime_params_lib
 from torax.geometry import pydantic_model as geometry_pydantic_model
 from torax.pedestal_model import pydantic_model as pedestal_pydantic_model
 from torax.sources import pydantic_model as source_pydantic_model
 from torax.stepper import pydantic_model as stepper_pydantic_model
-from torax.time_step_calculator import chi_time_step_calculator
-from torax.time_step_calculator import fixed_time_step_calculator
-from torax.time_step_calculator import time_step_calculator as time_step_calculator_lib
+from torax.time_step_calculator import pydantic_model as time_step_calculator_pydantic_model
 from torax.torax_pydantic import torax_pydantic
 from torax.transport_model import pydantic_model as transport_pydantic_model
 
@@ -56,18 +52,8 @@ def build_sim_from_config(
         sources: {...},
         transport: {...},
         stepper: {...},
-        time_step_calculator: {...},  # Optional
+        time_step_calculator: {...},
     }
-
-  See the following functions' docstrings to learn more about their expected
-  input config structures and what parameters are available to use.
-
-   -  `runtime_params`: geometry_pydantic_model.Geometry
-   -  `geometry`: `build_geometry_from_config()`
-   -  `sources`: `build_sources_from_config()`
-   -  `transport`: `build_transport_model_builder_from_config()`
-   -  `stepper`: stepper_pydantic_model.Stepper
-   -  `time_step_calculator`: `build_time_step_calculator_from_config()`
 
   To learn more about the Sim object and its components, see `sim.Sim`'s class
   docstring.
@@ -134,60 +120,8 @@ def build_sim_from_config(
       pedestal=pedestal_pydantic_model.Pedestal.from_dict(
           config['pedestal'] if 'pedestal' in config else {}
       ),
-      time_step_calculator=build_time_step_calculator_from_config(
+      time_step_calculator=time_step_calculator_pydantic_model.TimeStepCalculator.from_dict(
           config['time_step_calculator']
-      ),
+      ).time_step_calculator,
       file_restart=file_restart,
   )
-
-
-def build_time_step_calculator_from_config(
-    time_step_calculator_config: dict[str, Any] | str,
-) -> time_step_calculator_lib.TimeStepCalculator:
-  """Builds a `TimeStepCalculator` from the input config.
-
-  `TimeStepCalculator` calculates the dt for the time step and decides whether
-  the sim is done. The input config has one required key, `calculator_type`,
-  which must be one of the following values:
-
-   -  `fixed`: Maps to `FixedTimeStepCalculator`.
-   -  `chi`: Maps to `ChiTimeStepCalculator`.
-   -  `array`: Maps to `ArrayTimeStepCalculator`.
-
-  If the time-step calculator chosen has any constructor arguments, those can be
-  passed to the `init_kwargs` key in the input config:
-
-  .. code-block:: python
-
-    {
-        'calculator_type': 'array',
-        'init_kwargs': {
-            'array': [0, 0.1, 0.2, 0.4, 0.8],
-        }
-    }
-
-  Args:
-    time_step_calculator_config: Python dictionary configuring a
-      `TimeStepCalculator` with the structure shown above.
-
-  Returns:
-    A `TimeStepCalculator`.
-
-  Raises:
-    ValueError if the `calculator_type` is not one of the ones listed above.
-  """
-  if isinstance(time_step_calculator_config, str):
-    time_step_calculator_config = {
-        'calculator_type': time_step_calculator_config
-    }
-  else:
-    if 'calculator_type' not in time_step_calculator_config:
-      raise ValueError('calculator_type must be set in the input config.')
-    time_step_calculator_config = copy.deepcopy(time_step_calculator_config)
-  calculator_type = time_step_calculator_config.pop('calculator_type')
-  init_args = time_step_calculator_config.pop('init_kwargs', {})
-  if calculator_type == 'fixed':
-    return fixed_time_step_calculator.FixedTimeStepCalculator(**init_args)
-  elif calculator_type == 'chi':
-    return chi_time_step_calculator.ChiTimeStepCalculator(**init_args)
-  raise ValueError(f'Unknown calculator type: {calculator_type}')
